[WIP] CLI to anchor state roots
I created a new config module to handle configuration. It includes an
example configuration file which can be copied without the trailing
`.dist` ending and used as configuration. The config file is expected to
be in JSON format.
There is an accompanying `config.js` module that reads the config file
and also handles merging configuration from the file and command line
options.

In `src`, I created a new `cli` directory for CLI code and also added
the anchor as the first CLI. So far, it does nothing except get the
merged config from the command line and the config file and
`console.log`s it.

I guess the most interesting part is in the new `bin` directory. I
decided to use the npm package `commander`, as it takes a lot of the
boilerplate off our shoulders. I created the executable `mosaic` which
you can run from the command line.
It uses commander to specify commands, their arguments, and their
options. For anchoring, for example, the config file and the target
(origin or auxiliary) are mandatory arguments, whereas `address` and
`delay` are optional options to overwrite the configuration from the
file for that execution.
Commander automatically creates very good `-h, --help` output for
`mosaic` as well as the commands, in this case as an example `mosaic
anchor --help`.
I opted for all commands (anchor and facilitator) in a single `mosaic`
binary, as the meat should be in the `src/cli` directory. There would be
an option to specify the commands in `mosaic`, which in turn runs the
sub-commands from separate files. For example, `mosaic anchor` would run
the `mosaic-anchor` executable in the same directory. I deemed that
excessive for our use case, but it may becom an option in case that we
massively expand the number of commands or options of mosaic.

Example commands:
```
$ ./bin/mosaic --help
$ ./bin/mosaic -h
$ ./bin/mosaic --version
$ ./bin/mosaic anchor --help
$ ./bin/mosaic anchor ./config/config.json auxiliary
$ ./bin/mosaic anchor ./config/config.json auxiliary --address "0x123..." -d 12
```

Completely missing: sanity check of config file.

References #43