Add setup abstraction layer
* [x] code
* [ ] documentation
* [x] tests
* [ ] test assertions

**Initial Code for Setup**

I created a setup module with setup tools for gateways, organizations,
and anchors. They are still lacking JSDocs.

I started on integration tests for setup, but I am still lacking the
appropriate assertions

I am not 100% satisfied with the current architecture. Maybe someone has
an idea?

**Initialized GitBook**

Created the GitBook setup file and initial files to be filled in later.
I still want to add the full description for the setup part.