Facilitator.stake now returns nonce and msg hash
`Facilitator.stake()` now returns nonce and message hash. The events
are part of the web3 PromiEvent response and can be easily parsed from
the return object.

Refactored Facilitator. The stake function was too verbose. I extracted
some methods out to make the stake function more readable.

Updated integration tests to use different value token and base token.
Doing this also revealed the bug in the StakeHelper. The integration
tests now deploy two separate EIP20 tokens for base token and value
token. Those are now set on the Gateway and CoGateway as baes token and
value token.

The EIP20Gateway contract interact did not catch the case where the
possible reward is higher than the stake amount and the contract
reverts. I added the check as part of this ticket as it was a difficult
to track issue I faced with integration tests.

Added unit tests for facilitator and gateway. The facilitator stake unit
test now includes the response of nonce and message hash. The gateway
stake unit test now includes the check for the amount to be greater than
gas price times gas limit.

Added integration test for facilitator. I created a new file in the
integration test directory and let the facilitator stake with the
gateways that were deployd as part of the setup module integration test.
Because of the proper token setup on the gateways, I had to add a line
to the stake helper functional test where I set the value token (ref. #137)

Various linting errors fixed.

Fixes #136