StakeHelper now accepts value token in constructor
The stake helper sets both the value token and the simple token members
to the given simple token. This lead to errors in integration tests when
working on #136.

I propose this solution, where the argument is added at the end of the
constructor. This way, I would assume that KIT is not affected. I am
*guessing* that KIT does what I did in the integration test of #136:
they set `stakeHelper.valueToken` after initializing it.

Let's discuss if this is a viable solution or if we can offer something
else.

Fixes #137