Added return parameters to Facilitator::redeem()
I added the return parameters the same way as for stake.
I added a unit test that should assert that it works.

On the CoGateway contract interact, I added a check so that the amount
is always greater than gas price times gas limit. I added an
accompanying unit test for that.

I gave up on an integration test and thus I don't know if my change
actually works. I could not do an integration test for multiple reasons:

1. Overall, the integration tests are written in a hardly readable way.
2. I could not redeem, because:
   * There is no UtilityToken set up, only OSTPrime.
   * There is no minted UtilityToken to redeem on auxiliary (could not
     mint due to 3.)
   * I could not redeem as the chainOwner from OSTPrime (not sure why
     not)
3. I could not progress the stake to redeem again, as our test nodes
   don't support getting the proof.
4. Our test nodes do not support `debug` to attach and debug transaction
   failure.

So, we should test it with JLP once it is released.
Also, the integration tests need a massive overhaul (ref. #142).

All-in-all I am **very** concerned with the code quality of mosaic.js.